+++ R/transform-code.R
@@ -11,7 +11,7 @@
 transform_code <- function(path, fun, ..., dry) {
   if (is_plain_r_file(path) || is_rprofile_file(path)) {
     transform_utf8(path, fun = fun, ..., dry = dry)
-  } else if (is_rmd_file(path)) {
+  } else if (is_rmd_file(path) || is_qmd_file(path)) {
     transform_utf8(path,
       fun = partial(transform_mixed, transformer_fun = fun, filetype = "Rmd"),
       ..., dry = dry